Esm merging

+++ routes/routes.js
@@ -1,11 +1,14 @@
-const express = require("express");
+import express from "express";
+import { pool as connection, poolSkor as connectionSkor } from "../src/database.js";
+import crypto from "crypto";
+import bcrypt from "bcrypt";
+import dotenv from "dotenv";
+import axios from "axios";
+import he from "he";
+dotenv.config();
 const router = express.Router();
-const connection = require("../src/database");
-const crypto = require("crypto");
-const bcrypt = require("bcrypt");
-const request = require("sync-request");
-const SECRET_KEY = process.env.SECRET_KEY || "RED_AGATE_IS_RED"; // Harus disimpan di ENV
-const COOKIE_NAME = "red_agate";
+const SECRET_KEY = process.env.SECRET_KEY;
+const COOKIE_NAME = process.env.COOKIE_NAME;
 
 const monthMap = {
   0: "Januari",
@@ -22,23 +25,24 @@ const monthMap = {
   11: "Desember",
 };
 
-function generateAuthToken() {
+function generateAuthToken(perm = "score_viewer") {
   const timestamp = Date.now(); // Timestamp untuk mencegah replay attack
-  const data = `red_agate_is_good.${timestamp}`; // Format token
+  const data = `${perm}.${timestamp}`; // Format token
   const hmac = crypto
     .createHmac("sha256", SECRET_KEY)
     .update(data)
     .digest("hex"); // Buat signature HMAC
   return `${data}.${hmac}`; // Gabungkan data + signature
 }
-function verifyAuthToken(token) {
+function verifyAuthToken(req, token) {
   if (!token) return false;
 
   const parts = token.split(".");
   if (parts.length !== 3) return false;
 
-  const [prefix, timestamp, receivedHmac] = parts;
-  const data = `${prefix}.${timestamp}`;
+  const [perm, timestamp, receivedHmac] = parts;
+  const data = `${perm}.${timestamp}`;
+  req.session.RedAgateIsBeautiful=perm;
 
   // Cek apakah token sudah kedaluwarsa
   const tokenTime = parseInt(timestamp, 10);
@@ -58,11 +62,11 @@ function requireAuth(req, res, next) {
   const token = req.cookies[COOKIE_NAME];
 
   // Jika tidak ada token, redirect ke login
-  if (!token || !verifyAuthToken(token)) {
+  if (!token || !verifyAuthToken(req, token)) {
     res.clearCookie(COOKIE_NAME); // Hapus cookie jika tidak valid
     req.session.alertMessage = {
       type: "error",
-      message: "Sesi Anda telah kadaluarsa, silakan login kembali!",
+      message: "Sesi Anda telah kadaluarsa/Anda tidak memiliki akses, silakan login kembali!",
     };
     return res.redirect("/login");
   }
@@ -125,8 +129,26 @@ const getWBSCounter = (date) => {
   return counter;
 };
 
+
+
+
+
+
+
+
+
+
+
+
 // Main Router
 router.get("/export", requireAuth, async (req, res) => {
+  if(req.session.RedAgateIsBeautiful == 'score_viewer'){
+    req.session.alertMessage = {
+      type: "error",
+      message: "Anda tidak memiliki akses ke halaman ini!",
+    };
+    return res.redirect("/");
+  }
   connection.query(
     `
     SELECT udt, udtmonth, udtyear, situation, reporttitle, reportagent, 
@@ -165,14 +187,14 @@ router.get("/export", requireAuth, async (req, res) => {
 
 
 router.get("/about", requireAuth, async (req, res) => {
-  res.render("about", { title: "About | DEWANEV" });
+  res.render("about", { title: "About | DEWANEV", perm: req.session.RedAgateIsBeautiful });
 });
 
 router.get("/", requireAuth, async (req, res) => {
   const alertMessage = req.session.alertMessage;
   delete req.session.alertMessage;
 
-  res.render("index", { alertMessage, title: "Home | DEWANEV" });
+  res.render("index", { alertMessage, title: "Home | DEWANEV", perm: req.session.RedAgateIsBeautiful });
 });
 
 router.get("/login", (req, res) => {
@@ -195,12 +217,10 @@ router.get("/login", (req, res) => {
 
 router.post('/login', async (req, res) => {
   const { username, password } = req.body;
-  console.log('hello');
 
   try {
       // Fetch the first user from controlData
-      const result = await connection.query('SELECT * FROM controlData ORDER BY id LIMIT 1;');
-      console.log(result.rows);
+      const result = await connection.query('SELECT * FROM controlData WHERE username = $1', [username]);
       if (result.rows.length === 0) {
           req.session.alertMessage = {
               type: "error",
@@ -210,7 +230,7 @@ router.post('/login', async (req, res) => {
       }
 
       const user = result.rows[0];
-      console.log(user);
+
 
       if (user.username === username && bcrypt.compareSync(password, user.pwd)) {
           req.session.user = { username }; // Simpan di session
@@ -223,42 +243,59 @@ router.post('/login', async (req, res) => {
                           type: "error",
                           message: "Terjadi kesalahan saat hashing.",
                       };
-                      return res.render('login', { alertMessage: req.session.alertMessage, title: "Login | DEWANEV" });
+                      let alertMessage = req.session.alertMessage;
+                      delete req.session.alertMessage;
+                      return res.render('login', { alertMessage, title: "Login | DEWANEV" });
                   }
-                  const token = generateAuthToken(); // Buat token autentikasi
+                  const token = generateAuthToken(user.perm); // Buat token autentikasi
                   res.cookie(COOKIE_NAME, token, { httpOnly: true, maxAge: 1000 * 60 * 60 * 24 }); // Simpan cookie
-                  return res.redirect('/dashboard');
+                  delete req.session.alertMessage;
+                  return res.redirect('/');
               });
           });
           return;
       }
 
-      req.session.alertMessage = {
+      let alertMessage = {
           type: "error",
           message: "Username / Kata Sandi Salah!",
       };
-      res.render('login', { alertMessage: req.session.alertMessage, title: "Login | DEWANEV" });
+      res.render('login', { alertMessage, title: "Login | DEWANEV" });
 
   } catch (err) {
-      req.session.alertMessage = {
+      let alertMessage = {
           type: "error",
           message: "Terjadi Error! CodeError: " + err.message,
       };
-      res.render('login', { alertMessage: req.session.alertMessage, title: "Login | DEWANEV" });
+      res.render('login', { alertMessage, title: "Login | DEWANEV" });
   }
 });
 
 
 // Logout: Hapus session dan cookie
 router.get("/logout", (req, res) => {
-  res.clearCookie("user");
-  req.session.destroy(() => {
-    res.redirect("/login", { title: "Login | DEWANEV" });
+  res.clearCookie(COOKIE_NAME);
+  req.session.destroy((err) => {
+    if (err){
+      req.session.alertMessage({
+        type: "error",
+        message: "Tidak bisa log-out: "+err.message,
+      })
+      return res.redirect('/')
+    }
+    res.redirect("/login");
   });
 });
 
 // Lapor routers
 router.get("/lapor", requireAuth, (req, res) => {
+  if(req.session.RedAgateIsBeautiful == 'score_viewer'){
+    req.session.alertMessage = {
+      type: "error",
+      message: "Anda tidak memiliki akses ke halaman ini!",
+    };
+    return res.redirect("/");
+  }
   res.render("lapor", {
     title: "Lapor | DEWANEV",
     csrfToken: res.locals.csrfToken,
@@ -266,6 +303,13 @@ router.get("/lapor", requireAuth, (req, res) => {
 });
 
 router.post("/lapor", requireAuth, (req, res) => {
+  if(req.session.RedAgateIsBeautiful == 'score_viewer'){
+    req.session.alertMessage = {
+      type: "error",
+      message: "Anda tidak memiliki akses ke halaman ini!",
+    };
+    return res.redirect("/");
+  }
   let {
     reportTitle,
     knownTimeHappened,
@@ -334,6 +378,13 @@ router.post("/lapor", requireAuth, (req, res) => {
 
 // Edit routers
 router.get("/edit/:uid", requireAuth, (req, res) => {
+  if(req.session.RedAgateIsBeautiful == 'score_viewer'){
+    req.session.alertMessage = {
+      type: "error",
+      message: "Anda tidak memiliki akses ke halaman ini!",
+    };
+    return res.redirect("/");
+  }
   const uid = req.params.uid;
   connection.query(
     "SELECT * FROM LaporanTable WHERE uid = $1",
@@ -350,6 +401,13 @@ router.get("/edit/:uid", requireAuth, (req, res) => {
 });
 
 router.post("/edit/:uid", requireAuth, (req, res) => {
+  if(req.session.RedAgateIsBeautiful == 'score_viewer'){
+    req.session.alertMessage = {
+      type: "error",
+      message: "Anda tidak memiliki akses ke halaman ini!",
+    };
+    return res.redirect("/");
+  }
   const uid = req.params.uid;
   const {
     reportTitle,
@@ -413,6 +471,13 @@ router.post("/edit/:uid", requireAuth, (req, res) => {
   );
 });
 router.post("/verify", requireAuth, (req, res) => {
+  if(req.session.RedAgateIsBeautiful == 'score_viewer'){
+    req.session.alertMessage = {
+      type: "error",
+      message: "Anda tidak memiliki akses ke halaman ini!",
+    };
+    return res.redirect("/");
+  }
   const uid = req.body.uid;
 
   // Proses verifikasi di database...
@@ -445,6 +510,13 @@ router.post("/verify", requireAuth, (req, res) => {
 
 // Verifikasi routers
 router.post("/unverify", requireAuth, (req, res) => {
+  if(req.session.RedAgateIsBeautiful == 'score_viewer'){
+    req.session.alertMessage = {
+      type: "error",
+      message: "Anda tidak memiliki akses ke halaman ini!",
+    };
+    return res.redirect("/");
+  }
   const uid = req.body.uid;
 
   // Proses verifikasi di database...
@@ -479,6 +551,13 @@ router.post("/unverify", requireAuth, (req, res) => {
 
 // Delete routers
 router.post("/delete", requireAuth, (req, res) => {
+  if(req.session.RedAgateIsBeautiful == 'score_viewer'){
+    req.session.alertMessage = {
+      type: "error",
+      message: "Anda tidak memiliki akses ke halaman ini!",
+    };
+    return res.redirect("/");
+  }
   const { uid } = req.body;
   connection.query(
     "DELETE FROM LaporanTable WHERE uid = $1",
@@ -510,6 +589,13 @@ router.post("/delete", requireAuth, (req, res) => {
 
 // Dashboard routers
 router.get("/dashboard", requireAuth, async (req, res) => {
+  if(req.session.RedAgateIsBeautiful == 'score_viewer'){
+    req.session.alertMessage = {
+      type: "error",
+      message: "Anda tidak memiliki akses ke halaman ini!",
+    };
+    return res.redirect("/");
+  }
   const sortOption = req.query.sort || "latest"; // Default: terbaru
   let orderByClause = "ORDER BY uploadtime DESC"; // Default: terbaru
   
@@ -553,6 +639,7 @@ router.get("/dashboard", requireAuth, async (req, res) => {
     const results = await connection.query(dataQuery, [`%${search}%`, limit, offset]);
   
     res.render("dashboard", {
+      perm: req.session.RedAgateIsBeautiful,
       data: results.rows,
       totalPages,
       currentPage: page,
@@ -568,10 +655,161 @@ router.get("/dashboard", requireAuth, async (req, res) => {
     };
     res.redirect("/");
   }
-});  
+});
+
+
+// Skor Web
+
+// Route untuk menampilkan halaman skor
+router.get('/skor', requireAuth, async (req, res) => {
+  const { sortBy } = req.query; // Mendapatkan parameter query dari dropdown
+
+  let orderClause = 'ORDER BY nama'; // Default sorting (alfabet)
+  if (sortBy === 'score_asc') {
+    orderClause = 'ORDER BY skor ASC'; // Skor terendah
+  } else if (sortBy === 'score_desc') {
+    orderClause = 'ORDER BY skor DESC'; // Skor tertinggi
+  } else if (sortBy === 'name_desc') {
+    orderClause = 'ORDER BY nama DESC'; // Alfabet terbalik
+  }
+
+  try {
+    // Ambil data siswa berdasarkan status
+    const gdResult = await connectionSkor.query(`SELECT * FROM skor WHERE status = $1 ${orderClause}`, ['GD']);
+    const gdpResult = await connectionSkor.query(`SELECT * FROM skor WHERE status = $1 ${orderClause}`, ['GDP']);
+    const dppResult = await connectionSkor.query(`SELECT * FROM skor WHERE status = $1 ${orderClause}`, ['DPP']);
+
+    // Pisahkan data berdasarkan status
+    const gdStudents = gdResult.rows;
+    const gdpStudents = gdpResult.rows;
+    const dppStudents = dppResult.rows;
+
+    // Render halaman skor dengan data siswa
+    res.render('skor', { csrfToken: res.locals.csrfToken, gdStudents, gdpStudents, dppStudents, sortBy, title: "Skor | DEWANEV", perm: req.session.RedAgateIsBeautiful });
+  } catch (err) {
+    console.error('Error fetching data:', err);
+    res.status(500).send('Error fetching data');
+  }
+});
+
+
+router.post('/update-score-add/:id', requireAuth, async (req, res) => {
+  const { id } = req.params;
+  const { dropdown } = req.body;
+
+  axios.get(`https://api.github.com/gists/${process.env.GIST_ID}`, {headers: {'Authorization': `token ${process.env.GITHUB_TOKEN}`}})
+  .then(async response => {
+    const gist = response.data;
+    const fileContent = gist.files[process.env.GIST_FN].content
+
+    const penalties = JSON.parse(fileContent);
+
+    const result = await connectionSkor.query('SELECT * FROM skor WHERE uid = $1', [id]);
+    const student = result.rows[0];
+
+    if (student && penalties[dropdown]) {
+      let cskor = student.skor;
+      // Perbarui skor sesuai aksi
+
+      // Update skor di database
+      await connectionSkor.query('UPDATE skor SET skor = $1 WHERE uid = $2', [cskor+penalties[dropdown], id]);
+    }
+
+    // Kembali ke halaman skor dengan data yang diperbarui
+    res.redirect('/profile/' + id);
+  })
+  .catch (err => {
+    console.error('Error updating score:', err);
+    res.status(500).send('Error updating score');
+  });
+});
+
+function getScoreColor(score) {
+  if (score >= 25) return 'text-emerald-600'; // Hijau (30-25)
+  if (score >= 20) return 'text-lime-600'; // Hijau muda (20-24)
+  if (score >= 15) return 'text-yellow-600'; // Kuning (15-19)
+  if (score >= 10) return 'text-orange-600'; // Oranye (10-14)
+  if (score >= 5) return 'text-red-600'; // Oranye kemerahan (5-9)
+  return 'text-red-600'; // Merah (0-4)
+}
+
+
+
+router.get('/profile/:id', requireAuth, (req, res) => {
+  connectionSkor.query('SELECT * FROM skor WHERE uid = $1', [req.params.id], (err, resultsSkor) => {
+    connection.query('SELECT * FROM laporantable WHERE reportguilties ILIKE $1', [`%${resultsSkor.rows[0].nama}%`], (err, results) => {
+      axios.get(`https://api.github.com/gists/${process.env.GIST_ID}`, {headers: {'Authorization': `token ${process.env.GITHUB_TOKEN}`}})
+      .then(response => {
+        const gist = response.data;
+        const fileContent = gist.files[process.env.GIST_FN].content
+        res.render('profile', {
+          title: "Profile | DEWANEV",
+          student: resultsSkor.rows[0],
+          panduan: he.decode(fileContent),
+          getScoreColor,
+          perm: req.session.RedAgateIsBeautiful,
+          trackrecords: results.rows,
+          csrfToken: res.locals.csrfToken,
+        });
+        if (err) {
+          req.session.alertMessage = {
+            type: "error",
+            message: "Gagal mengambil data! CodeError: " + err.message,
+          };
+          return res.redirect('/skor');
+        }
+      })
+      .catch(error => {
+        console.error('Error fetching Gist:', error);
+      });
+    });
+    if (err) {
+      req.session.alertMessage = {
+        type: "error",
+        message: "Gagal mengambil data! CodeError: " + err.message,
+      };
+      return res.redirect('/skor');
+    }
+
+  });
+});
+
+
+// Route untuk memperbarui skor
+router.post('/update-score-sub/:id', requireAuth, async (req, res) => {
+  const { id } = req.params;
+  const { dropdown } = req.body;
+
+  axios.get(`https://api.github.com/gists/${process.env.GIST_ID}`, {headers: {'Authorization': `token ${process.env.GITHUB_TOKEN}`}})
+  .then(async response => {
+    const gist = response.data;
+    const fileContent = gist.files[process.env.GIST_FN].content
+
+    const penalties = JSON.parse(fileContent);
+
+    const result = await connectionSkor.query('SELECT * FROM skor WHERE uid = $1', [id]);
+    const student = result.rows[0];
+
+    if (student && penalties[dropdown]) {
+      let cskor = student.skor;
+      // Perbarui skor sesuai aksi
+
+      // Update skor di database
+      await connectionSkor.query('UPDATE skor SET skor = $1 WHERE uid = $2', [cskor-penalties[dropdown], id]);
+    }
+
+    // Kembali ke halaman skor dengan data yang diperbarui
+    res.redirect('/profile/' + id);
+  })
+  .catch (err => {
+    console.error('Error updating score:', err);
+    res.status(500).send('Error updating score');
+  });
+});
+
 
 router.use("/ada_sesuatu_yang_aneh", (req, res) => {
   res.render("error", { error: 500, title: "??? | DEWANEV", easter_egg: true });
 });
 
-module.exports = router;
+export default router;
